pyproject.toml
[tool.poetry]
name = "ihab"
version = "0.0.0"
description = "My project to create software similar to YNAB web for learning purposes"
authors = ["MrDynamo <mrdynamo@dynamiclab.org>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
Pillow = "^10.0.1"
pandas = "^2.1.1"
flask = "^3.0.0"
flask-cors = "^4.0.0"
flask-sqlalchemy = "^3.1.1"

[tool.poetry.dev-dependencies]
mkdocs-material = "^8.2.3"
pytest = "^6.2.1"
pytest-cov = "^2.11.0"
black = "^22.8"
prettier = "^0.0.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=ihab"
python_files = 'test_*'
python_classes = '*Tests'
python_functions = 'test_*'
testpaths = [
    "tests",
]
